Save token selection and amounts when going back
